working dir e2e fixture
