+++ tests/e2e/conftest.py
@@ -1,4 +1,5 @@
 import json
+import shutil
 import sys
 import tempfile
 from typing import List
@@ -6,12 +7,42 @@ from typing import Tuple
 from unittest.mock import patch
 
 import pytest
+from expected_download import _get_files_in_directory
 
 from ytdl_sub.cli.main import main
 from ytdl_sub.config.config_file import ConfigFile
 from ytdl_sub.subscriptions.subscription import Subscription
+from ytdl_sub.subscriptions.subscription_download import SubscriptionDownload
 from ytdl_sub.utils.file_handler import FileHandlerTransactionLog
+from ytdl_sub.utils.logger import Logger
 from ytdl_sub.utils.yaml import load_yaml
+
+logger = Logger.get("test")
+
+
+@pytest.fixture
+def working_directory() -> str:
+    """
+    Any time the working directory is used, ensure no files remain on cleaning it up
+    """
+    with tempfile.TemporaryDirectory() as temp_dir:
+
+        def _assert_working_directory_empty(self, is_error: bool):
+            files = [str(file_path) for file_path in _get_files_in_directory(temp_dir)]
+            num_files = len(files)
+            shutil.rmtree(temp_dir)
+
+            if not is_error:
+                if num_files > 0:
+                    logger.error("left-over files in working dir:\n%s", "\n".join(files))
+                assert num_files == 0
+
+        with patch.object(
+            SubscriptionDownload,
+            "_delete_working_directory",
+            new=_assert_working_directory_empty,
+        ):
+            yield temp_dir
 
 
 @pytest.fixture()
